Optimize fading

+++ src/client/rocktree.hpp
@@ -35,7 +35,7 @@ public:
 	bool is_buffering() const;
 	bool mark_for_buffering();
 
-	double draw(const shader_context& ctx, double current_time, const std::array<double, 8>& child_draw_time,
+	float draw(const shader_context& ctx, float current_time, const std::array<float, 8>& child_draw_time,
 	            const std::array<int, 8>& octant_mask);
 
 	static void buffer_queue(std::queue<node*> nodes);
@@ -51,7 +51,7 @@ private:
 	};
 
 	size_t last_draw_frame{0};
-	std::optional<double> draw_time_{};
+	std::optional<float> draw_time_{};
 	std::atomic<buffer_state> buffer_state_{buffer_state::unbuffered};
 
 	uint32_t epoch_{};
